add separate explorer service
connect to the websocket subscription and process blocks.
for the time being it's only a partial extraction of the API, as some
parts depend on the node's ledger.
the idea is to keep the config as much as possible the same as the node
(where it applies) more refactors may be needed to mimic the node's
config.